Share util functions
