+++ internal/builders/go/main_test.go
@@ -4,6 +4,7 @@ import (
 	"bufio"
 	"bytes"
 	"errors"
+	"fmt"
 	"io"
 	"os"
 	"os/exec"
@@ -11,17 +12,18 @@ import (
 	"regexp"
 	"testing"
 
-	"github.com/slsa-framework/slsa-github-generator/internal/builders/go/pkg"
-
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
+
+	"github.com/slsa-framework/slsa-github-generator/internal/builders/go/pkg"
+	"github.com/slsa-framework/slsa-github-generator/internal/utils"
 )
 
 func errCmp(e1, e2 error) bool {
 	return errors.Is(e1, e2) || errors.Is(e2, e1)
 }
 
-func Test_runVerify(t *testing.T) {
+func Test_runBuild(t *testing.T) {
 	t.Parallel()
 	tests := []struct {
 		subject    string
@@ -238,11 +240,6 @@ func Test_runVerify(t *testing.T) {
 			err:    pkg.ErrorInvalidEnvironmentVariable,
 		},
 		{
-			name:   "invalid main",
-			config: "./pkg/testdata/releaser-invalid-main.yml",
-			err:    pkg.ErrorInvalidDirectory,
-		},
-		{
 			name:   "invalid path",
 			config: "../pkg/testdata/releaser-invalid-main.yml",
 			err:    pkg.ErrorInvalidDirectory,
@@ -270,6 +267,7 @@ func Test_runVerify(t *testing.T) {
 			s := r.end()
 
 			if !errCmp(err, tt.err) {
+				fmt.Println(err, tt.err)
 				t.Errorf(cmp.Diff(err, tt.err, cmpopts.EquateErrors()))
 			}
 
@@ -391,12 +389,12 @@ func extract(lines string) ([]string, []string, string, string, error) {
 		return []string{}, []string{}, "", "", err
 	}
 
-	cmd, err := pkg.UnmarshallList(scmd)
+	cmd, err := utils.UnmarshallList(scmd)
 	if err != nil {
 		return []string{}, []string{}, "", "", err
 	}
 
-	env, err := pkg.UnmarshallList(senv)
+	env, err := utils.UnmarshallList(senv)
 	if err != nil {
 		return []string{}, []string{}, "", "", err
 	}
