+++ internal/utils/marshall.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package pkg
+package utils
 
 import (
 	"encoding/base64"
@@ -20,7 +20,26 @@ import (
 	"fmt"
 )
 
-func marshallToString(args interface{}) (string, error) {
+func UnmarshallList(arg string) ([]string, error) {
+	var res []string
+	// If argument is empty, return an empty list early,
+	// because `json.Unmarshal` would fail.
+	if arg == "" {
+		return res, nil
+	}
+
+	cs, err := base64.StdEncoding.DecodeString(arg)
+	if err != nil {
+		return res, fmt.Errorf("base64.StdEncoding.DecodeString: %w", err)
+	}
+
+	if err := json.Unmarshal(cs, &res); err != nil {
+		return []string{}, fmt.Errorf("json.Unmarshal: %w", err)
+	}
+	return res, nil
+}
+
+func MarshallToString(args interface{}) (string, error) {
 	jsonData, err := json.Marshal(args)
 	if err != nil {
 		return "", fmt.Errorf("json.Marshal: %w", err)
@@ -33,8 +52,8 @@ func marshallToString(args interface{}) (string, error) {
 	return encoded, nil
 }
 
-func marshallToBytes(args interface{}) ([]byte, error) {
-	encoded, err := marshallToString(args)
+func MarshallToBytes(args interface{}) ([]byte, error) {
+	encoded, err := MarshallToString(args)
 	if err != nil {
 		return []byte{}, nil
 	}
